requirements/base.txt
Django==4.1.7
djangorestframework==3.14.0
pydantic==1.10.5
serpy==0.3.1
pydantic[email]
email-validator
psycopg2-binary==2.9.5
Cerberus==1.3.4
drf-yasg==1.21.5
